# zstd: Faster decoding memcopy in asm