@klauspost 2022-05-09
It must really be hitting something it like. Maybe freeing up hyperthread capacity.