@lizthegrey 2022-05-12
I'm jealous :( hopefully Avo will support arm64 soon.